backend/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-cors-header
PILLOW
